requirements.txt
#python==3.6.2
certifi==2019.11.28
numpy==1.13.1
olefile==0.46
Pillow==4.2.1
scikit-learn==0.19.0
scipy==0.19.1
wincertstore==0.2
